Upgrade CDK, Node and Puppeteer
Since Node v14 is no longer supported as a lambda runtime, we upgrade all lambdas to Node v20.

Upgrade Puppeteer to v23 since v10 is not compatible with Node v20.

We utilise `@sparticuz/chromium` instead of `chrome-aws-lambda` since the latter is no longer
supported and therefore could not be used with Puppeteer >10.